Migrate settings and chat DOM tests to Playwright
